requirements.txt
python-dotenv
streamlit
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-huggingface
pypdf
faiss-cpu
transformers
sentence-transformers
nltk
torch